Move CompositionHelper into browser
